selftest: netfilter: add test case for stale udp connections
This commit adds a test case to verify that UDP conntrack entries
are not renewed by unidirectional trafic, to avoid cases when netfilter
DNAT rules are changed but the new packets keep matching the existing
conntrack rules that take precedence over existing netfilter rules.

The test case sets up specific DNAT rules and then establishes
a UDP connection. It then checks that packets reaches the expected
destination and changes the DNAT rule to reach a different destination.
Then it checks that after multiple requests, the sender is able to reach
the the new destination.

Based on test added by Florian Westphal in 37d220b.

Signed-off-by: Antonio Ojea <aojea@google.com>